Replace tagged pairs with records for ADTs
Instead of using a tagged pair to encode the value of an ADT produced by
a certain constructor, we now use Scheme records.

For each type constructor we emit a define-record-type form.

We are using the following convention: given a constructor Foo, the
uncurried function producing the value is Foo*, the predicate for such
constructor is Foo? and we reserve the use of Foo for the curried
constructor.