Optimize string operations by using a "cursor" (#235)
This optimizes many of the string operations by using a new abstraction
called _cursor_ that is much like a Scheme port but that works better for
our use cases. Cursor has an imperative API that reduces allocations
significantly compared to unconsing.

While removing the uses of `pstring-uncons-code-unit` I also found a bug
in both `pstring-index-of` and `pstring-last-index-of` where the match
wasn't found at all if the string and pattern shared a longer prefix.

This also changes how we decode code points from the UTF-16 bytevector:
we now replace lone surrogates with the Unicode Replacement Character.
The reason for this is that Scheme's `integer->char` fails for lone
surrogates and since purescm's `Char` is a Chez char we can more easily
treat code points as `Char`s. This way we can implement
`cursor-read-char` in terms of `cursor-read-code-point`.

Why not use Scheme ports for fast iteration over String? We could use
[`make-custom-textual-input-port`](https://scheme.com/tspl4/io.html#./io:s42)
but it's API assumes you want to allocate a Scheme string every time you
read from the port. This would work since we only ever need the first
character from the port, but is not optimal for this use case. We also
don't currently need the `port` abstraction for anything.